chore: move refreshReplicaConfiguration inside postgres.Instance
Closes #2225
